fixed app.js
